+++ extensions/src/org/pentaho/platform/datasource/MondrianDatasourceService.java
@@ -3,10 +3,14 @@ package org.pentaho.platform.datasource;
 import java.util.ArrayList;
 import java.util.List;
 
+import org.pentaho.metadata.model.Domain;
+import org.pentaho.metadata.repository.IMetadataDomainRepository;
 import org.pentaho.platform.api.datasource.GenericDatasourceServiceException;
 import org.pentaho.platform.api.datasource.IGenericDatasource;
 import org.pentaho.platform.api.datasource.IGenericDatasourceInfo;
 import org.pentaho.platform.api.datasource.IGenericDatasourceService;
+import org.pentaho.platform.api.engine.IAuthorizationPolicy;
+import org.pentaho.platform.api.engine.PentahoAccessControlException;
 import org.pentaho.platform.engine.core.system.PentahoSessionHolder;
 import org.pentaho.platform.engine.core.system.PentahoSystem;
 import org.pentaho.platform.plugin.action.mondrian.catalog.IMondrianCatalogService;
@@ -15,13 +19,19 @@ import org.pentaho.platform.plugin.action.mondrian.catalog.MondrianCatalog;
 public class MondrianDatasourceService implements IGenericDatasourceService{
   public static final String TYPE = "Analysis";
   IMondrianCatalogService mondrianCatalogService;
-
-  public MondrianDatasourceService() {
-    mondrianCatalogService = PentahoSystem.get(IMondrianCatalogService.class);
+  IMetadataDomainRepository metadataDomainRepository;
+  IAuthorizationPolicy policy;
+  ActionBasedSecurityService helper;
+  
+  public MondrianDatasourceService(IMondrianCatalogService mondrianCatalogService, IMetadataDomainRepository metadataDomainRepository, IAuthorizationPolicy policy) {
+    this.mondrianCatalogService = mondrianCatalogService;
+    this.metadataDomainRepository = metadataDomainRepository;
+    this.policy = policy;
+    helper = new ActionBasedSecurityService(policy);
   }
-
   @Override
-  public void add(IGenericDatasource datasource) throws GenericDatasourceServiceException {
+  public void add(IGenericDatasource datasource) throws GenericDatasourceServiceException, PentahoAccessControlException  {
+    helper.checkAdministratorAccess();
     if(datasource instanceof MondrianDatasource) {
       MondrianDatasource mondrianDatasource = (MondrianDatasource) datasource;
       mondrianCatalogService.addCatalog(mondrianDatasource.getDatasource(), false, PentahoSessionHolder.getSession());      
@@ -31,18 +41,21 @@ public class MondrianDatasourceService implements IGenericDatasourceService{
   }
 
   @Override
-  public MondrianDatasource get(String id) {
+  public MondrianDatasource get(String id) throws GenericDatasourceServiceException, PentahoAccessControlException  {
+    helper.checkAdministratorAccess();
     MondrianCatalog mondrianCatalog = mondrianCatalogService.getCatalog(id, PentahoSessionHolder.getSession());
     return new MondrianDatasource(mondrianCatalog, mondrianCatalog.getName(), TYPE);
   }
 
   @Override
-  public void remove(String id) throws GenericDatasourceServiceException {
+  public void remove(String id) throws GenericDatasourceServiceException, PentahoAccessControlException  {
+    helper.checkAdministratorAccess();
     mondrianCatalogService.removeCatalog(id, PentahoSessionHolder.getSession());
   }
 
   @Override
-  public void edit(IGenericDatasource datasource) throws GenericDatasourceServiceException {
+  public void edit(IGenericDatasource datasource) throws GenericDatasourceServiceException, PentahoAccessControlException  {
+    helper.checkAdministratorAccess();
     if(datasource instanceof MondrianDatasource) {
       MondrianDatasource mondrianDatasource = (MondrianDatasource) datasource;
       mondrianCatalogService.addCatalog(mondrianDatasource.getDatasource(), true, PentahoSessionHolder.getSession());
@@ -52,7 +65,8 @@ public class MondrianDatasourceService implements IGenericDatasourceService{
   }
 
   @Override
-  public List<IGenericDatasource> getAll() {
+  public List<IGenericDatasource> getAll() throws PentahoAccessControlException  {
+    helper.checkAdministratorAccess();
     List<IGenericDatasource> mondrianDatasourceList = new ArrayList<IGenericDatasource>();
     for(MondrianCatalog mondrianCatalog: mondrianCatalogService.listCatalogs(PentahoSessionHolder.getSession(), true)) {
       mondrianDatasourceList.add(new MondrianDatasource(mondrianCatalog, mondrianCatalog.getName(), TYPE));
@@ -61,10 +75,14 @@ public class MondrianDatasourceService implements IGenericDatasourceService{
   }
 
   @Override
-  public List<IGenericDatasourceInfo> getIds() {
+  public List<IGenericDatasourceInfo> getIds() throws PentahoAccessControlException  {
+    helper.checkAdministratorAccess();
     List<IGenericDatasourceInfo> datasourceInfoList = new ArrayList<IGenericDatasourceInfo>();
     for(MondrianCatalog mondrianCatalog: mondrianCatalogService.listCatalogs(PentahoSessionHolder.getSession(), true)) {
-      datasourceInfoList.add(new GenericDatasourceInfo(mondrianCatalog.getName(), TYPE));
+      Domain domain = metadataDomainRepository.getDomain(mondrianCatalog.getName());
+      if(domain == null) {
+        datasourceInfoList.add(new GenericDatasourceInfo(mondrianCatalog.getName(), TYPE));
+      }
     }
     return datasourceInfoList;
   }
